schedule indexing missing block on pending new block
