+++ services/blockchain-indexer/shared/indexer/utils/blockchainIndex.js
@@ -32,8 +32,6 @@ const { getLastIndexedBlock } = require('../lastIndexedBlock');
 const config = require('../../../config');
 
 const blocksTableSchema = require('../../database/schema/blocks');
-const { requestCoordinator } = require('../../utils/request');
-const { waitForCoordinatorReady } = require('./coordinator');
 
 const MYSQL_ENDPOINT = config.endpoints.mysql;
 
@@ -213,11 +211,6 @@ const updateTotalLockedAmounts = async (tokenIDLockedAmountChangeMap, dbTrx) =>
 		{ concurrency: Object.entries(tokenIDLockedAmountChangeMap).length },
 	);
 
-const scheduleMissingBlocksIndexing = async () => {
-	await waitForCoordinatorReady();
-	await requestCoordinator('scheduleMissingBlocksIndexing');
-};
-
 module.exports = {
 	reorderIndexBlocksQueueJobs,
 	activateReorderingMode,
@@ -227,5 +220,4 @@ module.exports = {
 	setLargestMissingBlockHeight,
 	getLargestMissingBlockHeight,
 	indexNewMissingBlock,
-	scheduleMissingBlocksIndexing,
 };
